+++ src/template/createTemplate.ts
@@ -12,5 +12,7 @@ export type TemplateReturn<T, C extends CRUDBase, U extends CRUDBase> = (
  * @returns Template for controllers
  */
 export const createTemplate = (data: (c: ControllerProps) => CreateTemplateData<unknown, CRUDBase, CRUDBase>) => {
-	return data as <T, C extends CRUDBase, U extends CRUDBase>(c: ControllerProps) => ControllerMethods<T, C, U>;
+	return data as <T, C extends CRUDBase, U extends CRUDBase, ISARR = true>(
+		c: ControllerProps
+	) => ControllerMethods<T, C, U, ISARR>;
 };