+++ src/utils/index.ts
@@ -17,18 +17,30 @@ export const fetcher = <T>(
 	data?: (object & CRUDBase) | string,
 	server?: string
 ): Promise<T> => {
-	const dataId = `${typeof data === 'string' ? data : data != null ? data.id : ''}`;
+	// Convert data from object | string to string | null
+	const dataId = typeof data === 'string' ? data : data != null ? data.id : null;
 
+	// If using GET, we cannot pass body using fetch. Make sure to delete id so we don't get unnecessary query parameter
 	if (typeof data === 'object' && method == 'GET') {
 		delete data.id;
 		if (Object.keys(data).length == 0) data = undefined;
 	}
 
+	/* Create url as:
+        /api/[base]/[parentUrl]/[dataId]/[url]
+        /api/[base]/[parentUrl]/[url]
+        /api/[base]/[url]/[dataId]
+        /api/[base]/[url]
+    */
+	const urlParams = method == 'GET' && data ? `?${new URLSearchParams(data as Record<string, string>)}` : '';
 	const url =
 		`/api/` +
 		(c.$base ? `${c.$base}/` : '') +
-		(c.$parentUrl ? `${c.$parentUrl}/${dataId}/${c.$url}` : `${c.$url}/${dataId}`) +
-		(method == 'GET' && data ? '?' + new URLSearchParams(data as Record<string, string>) : '');
+		(c.$parentUrl
+			? c.$parentUrl + (dataId ? `/${dataId}/` : '/') + c.$url + urlParams
+			: `${c.$url}/${dataId}` + urlParams);
+	// (c.$parentUrl ? `${c.$parentUrl}/${dataId}/${c.$url}` : `${c.$url}/${dataId}`) +
+	// (method == 'GET' && data ? '?' + new URLSearchParams(data as Record<string, string>) : '');
 
 	return fetch(url, {
 		method,